hc-python/app/requirements.txt
sqlalchemy-iris
testcontainers-iris
sentence-transformers
fastembed 
openai 
tiktoken
pandas
ipykernel
setuptools
Flask
python-dotenv